♻️ Split the types and `impl` blocks into a `detail` module to clarify the exposed public API
